rename folder and misc fix
