+++ app/src/main/java/com/example/mealplanner/ui/eventDetailActivity/DecisionFragment.kt
@@ -6,15 +6,11 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
 import com.example.mealplanner.databinding.FragmentDecisionBinding
-import com.example.mealplanner.databinding.FragmentGroupsBinding
-import com.example.mealplanner.ui.user.UserViewModel
 
 // chatGPT was used for parts of this file
 
 class DecisionFragment : Fragment() {
     private lateinit var binding: FragmentDecisionBinding
-    lateinit var restaurantName: String
-    var restaurantImageResourceId: Int = 0
     private lateinit var viewModel: EventDetailViewModel
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -26,7 +22,7 @@ class DecisionFragment : Fragment() {
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
+    ): View {
         binding = FragmentDecisionBinding.inflate(inflater, container, false)
 
         val root = binding.root
