+++ app/src/main/java/com/example/mealplanner/ui/login/LoginActivity.kt
@@ -15,7 +15,7 @@ import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import com.example.mealplanner.R
 import com.example.mealplanner.databinding.ActivityLoginBinding
-import com.example.mealplanner.ui.user.UserActivity
+import com.example.mealplanner.ui.userAcitivty.UserActivity
 
 
 class LoginActivity : AppCompatActivity() {
