+++ app/src/main/res/layout/fragment_groups.xml
@@ -5,7 +5,7 @@
     android:id="@+id/constraintLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.user.GroupsFragment">
+    tools:context=".ui.userAcitivty.GroupsFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recycler_view"
