+++ app/src/main/res/layout/fragment_profile.xml
@@ -5,7 +5,7 @@
     android:id="@+id/constraintLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.user.ProfileFragment">
+    tools:context=".ui.userAcitivty.ProfileFragment">
 
     <TextView
         android:id="@+id/name_lbl"
